Add Font and Flex typescript implementation
