+++ src/Flex/Flex.tsx
@@ -1,7 +1,9 @@
 import styled, {css} from "styled-components";
+import {TFlexProps} from "./Flex.types"
 
 
-export const Flex = (styled.div`
+
+export const Flex = (styled.div<TFlexProps>`
  display: flex;
  clear: both;
 
@@ -43,10 +45,10 @@ export const Flex = (styled.div`
  ${props => props.borderBox && css`box-sizing: border-box`}
 `);
 
-export const FlexRow = (styled(Flex)`
+export const FlexRow = (styled(Flex)<TFlexProps>`
   flex-direction: row;  
 `);
 
-export const FlexColumn = (styled(Flex)`
+export const FlexColumn = (styled(Flex)<TFlexProps>`
   flex-direction: column;
 `);
