+++ src/Font/Font.types.ts
@@ -1,0 +1,12 @@
+export interface TFontProps {
+    color?: string,
+    noMargin?: boolean,
+    light?: boolean,
+    thin?: boolean,
+    bold?: boolean,
+    marginRight?: string,
+    marginLeft?: string,
+    marginTop?: string,
+    marginBottom?: string,
+    noWrap?: boolean
+}
