--- src/components/library/flex/story/flex.stories.js
@@ -1,9 +1,0 @@
-import React from "react";
-import {DiceColumn} from "./dicecolumn";
-import {DiceRow} from "./dicerow";
-
-
-export default {title: "Flex"};
-
-export const Column = () => <DiceColumn/>;
-export const Row = () => <DiceRow/>;
