+++ src/index.ts
@@ -1,3 +1,11 @@
 import {Button} from "./Button/Button"
+import {Flex, FlexColumn, FlexRow} from "./Flex/Flex"
 
-export {Button}
+import {Caption, Headline, Paragraph, Subheadline, Text, Tiny} from "./Font/Font"
+
+
+export {
+    Button,
+    Flex, FlexColumn, FlexRow,
+    Caption, Headline, Paragraph, Subheadline, Text, Tiny
+}
